// frontend/src/api/generated/model/tagKey.ts
export type TagKey = (typeof TagKey)[keyof typeof TagKey];

export const TagKey = {
  fiction: "fiction",
  non_fiction: "non_fiction",
  reference: "reference",
  textbook: "textbook",
  anthology: "anthology",
  comics: "comics",
  manga: "manga",
  play: "play",
  essays: "essays",
  picture_book: "picture_book",
  adventure: "adventure",
  classic: "classic",
  contemporary_fiction: "contemporary_fiction",
  crime: "crime",
  detective: "detective",
  dystopian: "dystopian",
  epic_fantasy: "epic_fantasy",
  fairy_tales: "fairy_tales",
  fantasy: "fantasy",
  folklore: "folklore",
  gothic: "gothic",
  graphic_novel: "graphic_novel",
  historical_fiction: "historical_fiction",
  horror: "horror",
  humour: "humour",
  literary_fiction: "literary_fiction",
  magical_realism: "magical_realism",
  mystery: "mystery",
  mythology: "mythology",
  noir: "noir",
  paranormal: "paranormal",
  poetry: "poetry",
  post_apocalyptic: "post_apocalyptic",
  romance: "romance",
  satire: "satire",
  science_fiction: "science_fiction",
  short_stories: "short_stories",
  space_opera: "space_opera",
  speculative_fiction: "speculative_fiction",
  spy_fiction: "spy_fiction",
  steampunk: "steampunk",
  suspense: "suspense",
  thriller: "thriller",
  urban_fantasy: "urban_fantasy",
  war: "war",
  western: "western",
  anthropology: "anthropology",
  archaeology: "archaeology",
  architecture: "architecture",
  art: "art",
  astronomy: "astronomy",
  autobiography: "autobiography",
  biography: "biography",
  biology: "biology",
  business: "business",
  chemistry: "chemistry",
  computing: "computing",
  cooking: "cooking",
  design: "design",
  diaries_and_letters: "diaries_and_letters",
  economics: "economics",
  education: "education",
  environment: "environment",
  ethics: "ethics",
  feminism: "feminism",
  film_and_tv: "film_and_tv",
  finance: "finance",
  gardening: "gardening",
  geography: "geography",
  health_and_fitness: "health_and_fitness",
  history: "history",
  journalism: "journalism",
  language: "language",
  law: "law",
  linguistics: "linguistics",
  mathematics: "mathematics",
  medicine: "medicine",
  memoir: "memoir",
  music: "music",
  nature: "nature",
  parenting: "parenting",
  philosophy: "philosophy",
  photography: "photography",
  physics: "physics",
  politics: "politics",
  popular_science: "popular_science",
  psychology: "psychology",
  religion: "religion",
  science: "science",
  self_help: "self_help",
  sociology: "sociology",
  sports: "sports",
  technology: "technology",
  theatre: "theatre",
  travel: "travel",
  true_crime: "true_crime",
  urbanism: "urbanism",
  wine_and_drink: "wine_and_drink",
  baby_and_toddler: "baby_and_toddler",
  children: "children",
  early_reader: "early_reader",
  middle_grade: "middle_grade",
  young_adult: "young_adult",
  new_adult: "new_adult",
  adult: "adult",
} as const;